reorganized functions
